+++ packages/compiler/src/template/pipeline/src/ingest.ts
@@ -318,12 +318,10 @@ function ingestElement(unit: ViewCompilationUnit, element: t.Element): void {
 
   // We want to ensure that the controlCreateOp is after the ops that create the element
   const fieldInput = element.inputs.find(
-    (input) =>
-      (input.name === 'field' || input.name === 'formField') &&
-      input.type === e.BindingType.Property,
+    (input) => input.name === 'formField' && input.type === e.BindingType.Property,
   );
   if (fieldInput) {
-    // If the input name is 'field', this could be a form control binding which requires a
+    // If the input name is 'formField', this could be a form control binding which requires a
     // `ControlCreateOp` to properly initialize.
     unit.create.push(ir.createControlCreateOp(fieldInput.sourceSpan));
   }
