+++ packages/compiler/src/template/pipeline/src/phases/binding_specialization.ts
@@ -127,7 +127,7 @@ export function specializeBindings(job: CompilationJob): void {
                 op.sourceSpan,
               ),
             );
-          } else if (op.name === 'field' || op.name === 'formField') {
+          } else if (op.name === 'formField') {
             ir.OpList.replace<ir.UpdateOp>(op, ir.createControlOp(op));
           } else {
             ir.OpList.replace<ir.UpdateOp>(
