+++ packages/core/test/acceptance/create_component_spec.ts
@@ -6,6 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
+import {ChangeDetectionStrategy} from '@angular/compiler';
 import {
   Component,
   createComponent,
@@ -36,7 +37,6 @@ import {
 } from '../../src/core';
 import {stringifyForError} from '../../src/render3/util/stringify_utils';
 import {TestBed} from '../../testing';
-import {ChangeDetectionStrategy} from '@angular/compiler';
 
 describe('createComponent', () => {
   it('should create an instance of a standalone component', () => {
@@ -1139,14 +1139,14 @@ describe('createComponent', () => {
       expect(hostElement.textContent).toBe('Value: 20');
     });
 
-    it(`should support input bindings named 'field'`, () => {
-      // Angular has specialized support for binding to form controls (e.g. `[field]="field"`).
+    it(`should support input bindings named 'formField'`, () => {
+      // Angular has specialized support for binding to form controls (e.g. `[formField]="field"`).
       // This test ensures that dynamic input bindings can still target arbitrary inputs with the
       // same name.
 
       @Component({template: ''})
       class RootComp {
-        @Input() field = '';
+        @Input() formField = '';
       }
 
       const hostElement = document.createElement('div');
@@ -1156,15 +1156,15 @@ describe('createComponent', () => {
       const ref = createComponent(RootComp, {
         hostElement,
         environmentInjector,
-        bindings: [inputBinding('field', valueSignal)],
+        bindings: [inputBinding('formField', valueSignal)],
       });
 
       ref.changeDetectorRef.detectChanges();
-      expect(ref.instance.field).toBe('hello');
+      expect(ref.instance.formField).toBe('hello');
 
       valueSignal.set('goodbye');
       ref.changeDetectorRef.detectChanges();
-      expect(ref.instance.field).toBe('goodbye');
+      expect(ref.instance.formField).toBe('goodbye');
     });
   });
