+++ packages/forms/signals/public_api.ts
@@ -13,7 +13,6 @@
  */
 export * from './src/api/control';
 export * from './src/api/di';
-export * from './src/api/field_directive';
 export * from './src/api/form_field_directive';
 export * from './src/api/rules';
 export * from './src/api/rules/debounce';
