+++ packages/forms/signals/src/api/di.ts
@@ -8,7 +8,6 @@
 
 import {type Provider} from '@angular/core';
 import {SIGNAL_FORMS_CONFIG} from '../field/di';
-import type {Field} from './field_directive';
 import type {FormField} from './form_field_directive';
 
 /**
@@ -18,7 +17,7 @@ import type {FormField} from './form_field_directive';
  */
 export interface SignalFormsConfig {
   /** A map of CSS class names to predicate functions that determine when to apply them. */
-  classes?: {[className: string]: (state: Field<unknown> | FormField<unknown>) => boolean};
+  classes?: {[className: string]: (state: FormField<unknown>) => boolean};
 }
 
 /**
